+++ software/4/SAX/screen.h
@@ -23,7 +23,7 @@ class screen
     void draw_title_value(String title, char value, int N, int * values ,int sub_value = 0);
     void draw_title_value(String title, String * value, int N, int * values ,int sub_value = 0);
     void draw_single_value(String value);
-    void draw_standby_screen(int & _octave, int & _transpose, byte _global_mode, bool _delta_mode, byte _joystick, byte _joystick_bis, unsigned int tempo);
+    void draw_standby_screen(int8_t & _octave, int8_t & _transpose, byte _global_mode, bool _delta_mode, byte _joystick, byte _joystick_bis, unsigned int tempo);
     void draw_progress_bar(byte value);
     void draw_low_batt();
